+++ src/hgraph/_operators/_to_table.py
@@ -5,7 +5,8 @@ from operator import le
 from typing import TypeVar, Callable, Any
 
 from hgraph._operators._operators import operator
-from hgraph._runtime import GlobalState, EvaluationClock
+from hgraph._runtime._global_state import GlobalState
+from hgraph._runtime._evaluation_clock import EvaluationClock
 from hgraph._types import TIME_SERIES_TYPE, CompoundScalar, TS, OUT, DEFAULT
 from hgraph._types._scalar_type_meta_data import HgTupleFixedScalarType, HgTupleCollectionScalarType
 from hgraph._wiring._wiring_errors import CustomMessageWiringError