+++ src/hgraph/_wiring/_switch.py
@@ -1,16 +1,15 @@
 from inspect import isfunction
 from pathlib import Path
-from typing import Callable, Optional, cast, Any
+from typing import Callable, Optional, cast
 
 from frozendict import frozendict
 
-from hgraph._types._scalar_types import SCALAR, STATE, DEFAULT
-from hgraph._types._type_meta_data import HgTypeMetaData
+from hgraph._types._scalar_types import SCALAR, STATE
 from hgraph._types._time_series_meta_data import HgTimeSeriesTypeMetaData
 from hgraph._types._time_series_types import TIME_SERIES_TYPE
 from hgraph._types._ts_meta_data import HgTSTypeMetaData
 from hgraph._types._ts_type import TS
-from hgraph._wiring._markers import _Marker
+from hgraph._types._type_meta_data import HgTypeMetaData
 from hgraph._wiring._source_code_details import SourceCodeDetails
 from hgraph._wiring._wiring_context import WiringContext
 from hgraph._wiring._wiring_errors import CustomMessageWiringError
@@ -18,8 +17,8 @@ from hgraph._wiring._wiring_node_class._switch_wiring_node import SwitchWiringSi
 from hgraph._wiring._wiring_node_class._wiring_node_class import WiringNodeClass, extract_kwargs
 from hgraph._wiring._wiring_node_signature import WiringNodeSignature
 from hgraph._wiring._wiring_node_signature import WiringNodeType
-from hgraph._wiring._wiring_utils import as_reference, wire_nested_graph
 from hgraph._wiring._wiring_port import WiringPort
+from hgraph._wiring._wiring_utils import as_reference, wire_nested_graph
 
 __all__ = ("switch_",)