+++ src/hgraph/_wiring/_wiring_node_class/_mesh_wiring_node.py
@@ -46,5 +46,6 @@ class MeshWiringNodeClass(BaseWiringNodeClass):
             output_node_id=output_node_id,
             multiplexed_args=self.signature.multiplexed_args,
             key_arg=self.signature.key_arg,
+            key_tp=self.signature.key_tp,
             context_path=self.signature.context_path,
         )